+++ CHANGELOG.md
@@ -11,6 +11,7 @@ You must updated your config when updating from 0.1.13 to 1.0.0 (or later versio
 ## 1.0.2 (2021-03-22)
 ### Changes
 * Fixed a problem where the start command was not being sent to the backend platform
+* Fixed some small logging display errors
 * Optimised some code for the television service